try to fix one instance of Russian "ne"
Writing "ne" with something other than verbs is a complex subject.
One has to rely on Rozental rules, as well as on one's deep feeling
of the language.  One rule and one intuition says that, in a predicate
that can be accentuated, akin to Pushkin's "obmanut' menya ne trudno",
or Bulgakov's "eto ochen' ne prosto", "ne" should be written separately.